Moved mapper functions into their own file
