+++ src/Main/HtmlMapper/ParagraphElements/Html_ParagraphElements.ts
@@ -1,19 +1,18 @@
-import { ParagraphElement } from "../../InterfaceLayer/ParagraphElements"
-import { Html_Image, IHtml_Image } from "./Html_Image/Html_Image"
-import { Html_TextRun, IHtml_TextRun } from "./Html_TextRun/Html_TextRun"
+import { IHtml_Image } from "./Html_Image/Html_Image"
+import { IHtml_TextRun } from "./Html_TextRun/Html_TextRun"
+
 
 export interface IHtml_ParagraphElementType<A extends ElementTypes> {
     type: A
 }
-type ElementTypes = "html_image" | "html_textRun" 
+
+const HtmlElementTypes = {
+    html_image: "html_image",
+    html_textRun: "html_textRun"
+} as const
+
+type ElementTypes = typeof HtmlElementTypes[keyof typeof HtmlElementTypes]
 
 export type IImageElement = IHtml_Image & IHtml_ParagraphElementType<"html_image"> 
 export type ITextElement = IHtml_TextRun & IHtml_ParagraphElementType<"html_textRun">
 export type IHtml_ParagragraphElement = IImageElement | ITextElement
-
-
-export class Html_ParagraphElement{
-    static from(element:ParagraphElement):IHtml_ParagragraphElement{
-        return element.type == "image" ? Html_Image.from(element) : Html_TextRun.from(element)
-    }
-}
